# gin_helloworld/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8090",
    "basePath": "/api/v1",
    "paths": {
        "/info": {
            "get": {
                "description": "各种get参数",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试"
                ],
                "summary": "测试query参数获取",
                "parameters": [
                    {
                        "type": "string",
                        "description": "测试.ShouldBindQuery",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.ShouldBindQuery",
                        "name": "pwd",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.Param",
                        "name": "param2",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.Query",
                        "name": "param3",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.DefaultQuery",
                        "name": "param4",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.GetQuery",
                        "name": "param5",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"msg\": \"who are you\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/jsont": {
            "post": {
                "description": "各种get参数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试"
                ],
                "summary": "测试json参数获取",
                "parameters": [
                    {
                        "description": "测试.BindJSON",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Registjson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Respose"
                        }
                    },
                    "500": {
                        "description": "{\"msg\": \"error\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/regin": {
            "post": {
                "description": "各种get参数",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "测试"
                ],
                "summary": "测试form参数获取",
                "parameters": [
                    {
                        "type": "string",
                        "description": "测试.ShouldBind",
                        "name": "uid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.ShouldBind",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.PostForm",
                        "name": "param2",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.DefaultPostForm",
                        "name": "param3",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "测试.GetPostForm",
                        "name": "param4",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"msg\": \"ERROR\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Registjson": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "main.Respose": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}